feat: add more text to the about section
